chore: remove pkg/util/common.go
