+++ comp/logs/agent/agentimpl/agent.go
@@ -46,8 +46,8 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/logs/tailers"
 	"github.com/DataDog/datadog-agent/pkg/remoteconfig/state"
 	"github.com/DataDog/datadog-agent/pkg/status/health"
-	"github.com/DataDog/datadog-agent/pkg/util"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+	"github.com/DataDog/datadog-agent/pkg/util/goroutinesdump"
 	"github.com/DataDog/datadog-agent/pkg/util/optional"
 	"github.com/DataDog/datadog-agent/pkg/util/startstop"
 )
@@ -318,7 +318,7 @@ func (a *logAgent) stop(context.Context) error {
 		case <-c:
 		case <-timeout.C:
 			a.log.Warn("Force close of the Logs Agent, dumping the Go routines.")
-			if stack, err := util.GetGoRoutinesDump(); err != nil {
+			if stack, err := goroutinesdump.Get(); err != nil {
 				a.log.Warnf("can't get the Go routines dump: %s\n", err)
 			} else {
 				a.log.Warn(stack)
